Use int for read commands and unsigned int for their generated keys

+++ src/include/io.h
@@ -6,7 +6,7 @@
 
 int read_in_bf(size_t *);
 
-int read_ws_command_char(FILE *, size_t *);
+int read_ws_command_char(FILE *, int *);
 
 int tokenize_ws_raw(FILE *, size_t **, size_t *, size_t *);
